EDS-75 Dispatch auto-select with an <a href="">
https://issues.shibboleth.net/jira/browse/EDS-75

location.href=<address> causes (at least) Firefox to sent a 302 and
appears that Firefox then caches this.  Opinions differ as to whether
it is within it's right but "it is what it is".

The checkin stops using this technique and instead builds an <a>, adds
it to the page and then clicks it.  This causes (at least firefox to
send) a GET to the address and is what the "previously used icons"
thing does as well - and this has never been seen to cache.